# server/src/api-docs/swagger.json
{
    "openapi": "3.0.1",
    
    "info": 
    {
      "title": "La Bonne Alternance",
      "description": "Cherchez des formations et des emplois en apprentissages",
      "version": "V1.0",
      "contact": "equipelba@pole-emploi.fr"
    },

    "servers": [
        {
            "url": "https://idea.apprentissage.beta.gouv.fr"
        },
        {
            "url": "https://idea-recette.apprentissage.beta.gouv.fr"
        }
    ],

    "paths": 
    {        
        "/api/V1/formations": 
        {
            "get": 
            {
                "description": "Rechercher des formations en apprentissage pour un métier ou un ensemble de métiers autour d'un point géographique",

                "produces": [
                    "application/xml",
                    "application/json"
                ],

                "parameters": [
                    {
                        "$ref": "#/parameters/romes"
                    },
                    {
                        "$ref": "#/parameters/romeDomain"
                    },
                    {
                        "$ref": "#/parameters/latitude"                        
                    },
                    {
                        "$ref": "#/parameters/longitude"                        
                    },
                    {
                        "$ref": "#/parameters/radius"                        
                    },
                    {
                        "$ref": "#/parameters/diploma"
                    }
                    
                ],

                "responses": {
                    "200": 
                    {
                        "description": "Un tableau contenant la liste des formations correspondants aux critères transmis en paramètre de la requête. Le tableau peut être vide si aucune formation ne correspond.",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/definitions/Formations"
                                }
                            }
                        }
                    },
                    
                    "400": 
                    {
                        "$ref": "#/responses/WrongParameters"
                    }
                }            
            }
        },

        "/api/V1/formationsParRegion": 
        {
            "get": 
            {
                "description": "Rechercher des formations en apprentissage dans un département ou dans une région ou dans la France entière pour un métier ou un ensemble de métiers",
                
                "parameters": [
                    {
                        "$ref": "#/parameters/regionRomes"                        
                    },
                    {
                        "$ref": "#/parameters/regionRomeDomain"                        
                    },
                    {
                        "$ref": "#/parameters/region"                        
                    },
                    {
                        "$ref": "#/parameters/departement"                        
                    },
                    {
                        "$ref": "#/parameters/diploma"
                    }
                ],

                "responses": {
                    "200": 
                    {
                        "description": "Un tableau contenant la liste des formations correspondants aux critères transmis en paramètre de la requête. Le tableau peut être vide si aucune formation ne correspond.",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/definitions/Formations"
                                }
                            }
                        }
                    },
                    
                    "400": 
                    {
                        "$ref": "#/responses/WrongParameters"
                    }
                }            
            }
        },

        "/api/V1/jobs": 
        {
            "get": 
            {
                "description": "Rechercher les entreprises recrutant ou susceptibles de recruter en apprentissage référencées par les APIs de Pôle emploi, pour un métier ou un ensemble de métiers, autour d'un point géographique",
                
                "parameters": [
                    {
                        "$ref": "#/parameters/requiredRomes"                        
                    },
                    {
                        "$ref": "#/parameters/latitude"                        
                    },
                    {
                        "$ref": "#/parameters/longitude"                        
                    },
                    {
                        "$ref": "#/parameters/radius"                        
                    },
                    {
                        "$ref": "#/parameters/insee"
                    },
                    {
                        "$ref": "#/parameters/jobSources"
                    }
                ],

                "responses": {
                    "200": 
                    {
                        "description": "tableau contenant la liste des formations correspondants aux critères transmis en paramètre de la requête. Le tableau peut être vide si aucune formation ne correspond.",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/definitions/Jobs"
                                }
                            }
                        }
                    },
                    
                    "400": 
                    {
                        "$ref": "#/responses/WrongParameters"
                    }
                }            
            }
        },

        "/api/V1/jobsEtFormations": 
        {
            "get": 
            {
                "description": "Rechercher des formations en apprentissage et des entreprises recrutant ou susceptibles de recruter en apprentissage référencées par les APIs de Pôle emploi, pour un métier ou un ensemble de métiers, autour d'un point géographique",
                "parameters": [
                    {
                        "$ref": "#/parameters/requiredRomes"                        
                    },
                    {
                        "$ref": "#/parameters/latitude"                        
                    },
                    {
                        "$ref": "#/parameters/longitude"                        
                    },
                    {
                        "$ref": "#/parameters/radius"                        
                    },
                    {
                        "$ref": "#/parameters/diploma"
                    },
                    {
                        "$ref": "#/parameters/insee"
                    },
                    {
                        "$ref": "#/parameters/sources"
                    }
                ],

                "responses": {
                    "200": 
                    {
                        "description": "Un objet contenant la liste des formations et les opportunités d'emploi correspondants aux critères transmis en paramètre de la requête.",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "type": "object",
                                    "properties":{
                                        "formations":
                                        {
                                            "$ref": "#/definitions/Formations"
                                        },
                                        "jobs":
                                        {
                                            "$ref": "#/definitions/Jobs"
                                        }
                                    }
                                    
                                }
                            }
                        }
                    },
                    
                    "400": 
                    {
                        "$ref": "#/responses/WrongParameters"
                    }
                }            
            }
        }

    },
 
    "responses":
    {
        "WrongParameters" :
        {
            "description": "Mauvais paramètres",
            "content":
            {
                "application/json":
                {
                    "schema":
                    {
                        "$ref": "#/definitions/Error"
                    }
                }
            }
        }
    },
    
    "parameters":
    {
        "romes":
        {
            "name": "romes",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Une liste de codes ROME séparés par des virgules correspondant au(x) métier(s) recherché(s). Maximum 9.<br />rome et romeDomain sont incompatibles.<br/><strong>Au moins un des deux doit être renseigné.</strong>",
            "example": "F1603,I1308"
        },

        "regionRomes":
        {
            "name": "romes",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Une liste de codes ROME séparés par des virgules correspondant au(x) métier(s) recherché(s). Maximum 9.<br />rome et romeDomain sont incompatibles.<br /><strong>Au moins un des deux doit être renseigné dans le cas d'une recherche France entière.</strong>",
            "example": "F1603,I1308"
        },

        "requiredRomes":
        {
            "name": "romes",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Une liste de codes ROME séparés par des virgules correspondant au(x) métier(s) recherché(s). Maximum 9.",
            "example": "F1603,I1308"
        },

        "romeDomain":
        {
            "name": "romeDomain",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Un domaine ROME (1 lettre et deux chiffres) ou un grand domaine ROME (1 lettre). <br />rome et romeDomain sont incompatibles.<br /><strong>Au moins un des deux doit être renseigné.</strong>",
            "example": "F ou I13"
        },

        "regionRomeDomain":
        {
            "name": "romeDomain",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Un domaine ROME (1 lettre et deux chiffres) ou un grand domaine ROME (1 lettre).<br />rome et romeDomain sont incompatibles.<br /><strong>Au moins un des deux doit être renseigné dans le cas d'une recherche France entière.</strong>",
            "example": "F ou I13"
        },

        "region":
        {
            "name": "region",
            "in": "query",
            "required": false,
            "schema":{
                "type": "string",
                "enum": ["01", "02", "03", "04", "06", "11", "24", "27", "28", "32", "44", "52", "53", "75", "76", "84", "93", "94"]
            },
            "description": "Le code Insee d'une région française.<br />region et departement sont incompatibles.<br /><strong>Si aucun des deux n'est renseigné la recherche portera sur France entière.</strong><br /><strong>France entière ne peut être utilisé sans rome ou romeDomain.</strong><br /><ul><li>84 : Auvergne-Rhône-Alpes</li><li>27 : Bourgogne-Franche-Comté</li><li>53 : Bretagne</li><li>24 : Centre-Val de Loire</li><li>94 : Corse</li><li>44 : Grand Est</li><li>32 : Hauts-de-France</li><li>11 : Île-de-France</li><li>28 : Normandie</li><li>75 : Nouvelle-Aquitaine</li><li>76 : Occitanie</li><li>52 : Pays de la Loire</li><li>93 : Provence-Alpes-Côte d'Azur</li><li>01 : Guadeloupe</li><li>02 : Martinique</li><li>03 : Guyane</li><li>04 : La Réunion</li><li>06 : Mayotte</li></ul>",
            "example": "84"
        },

        "departement":
        {
            "name": "departement",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Le numéro à 2 chiffres d'un département français ou à 3 chiffres pour un département d'outre-mer.<br />region et departement sont incompatibles.<br /><strong>Si aucun des deux n'est renseigné la recherche portera sur France entière.</strong><br /><strong>France entière ne peut être utilisé sans rome ou romeDomain.</strong>",
            "example": "09 ou 974"
        },

        "insee":
        {
            "name": "insee",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Le code Insee de la commune de recherche. Ce doit être un nombre à cinq chiffres.",
            "example": "94071"
        },

        "latitude":
        {
            "name": "latitude",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "La latitude du centre de recherche",
            "example": "48.845"
        },

        "longitude":
        {
            "name": "longitude",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "La longitude du centre de recherche",
            "example": "2.3752"
        },

        "radius":
        {
            "name": "radius",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "Le rayon de recherche en kilomètres", 
            "example": 30,
            "default": 30
        },

        "diploma":
        {
            "name": "diploma",
            "in": "query",
            "required": false,
            "schema":{
                "type": "string",
                "enum": ["3 (CAP...)", "4 (Bac...)", "5 (BTS, DUT...)", "6 (Licence...)", "7 (Master, titre ingénieur...)"]
            },
            "description": "Le niveau de diplôme requis. Si précisé, doit contenir une seule des valeurs autorisées", 
            "example": "3 (CAP...)"
        },

        "sources":
        {
            "name": "sources",
            "in": "query",
            "required": false,
            "explode": false,
            "schema": {
                "type": "array",
                "minItems": 1,
                "collectionFormat": "csv",
                "default": "formations,lba,lbb,offres",
                "items":
                    {
                        "type": "string",
                        "enum": ["formations", "lba", "lbb", "offres"]                            
                    }
            },
            "description": "Les sources de données dans lesquelles la recherche aura lieu. Ces sources doivent être séparées par des virgules", 
            "example": "formations,offres"
        },

        "jobSources":
        {
            "name": "sources",
            "in": "query",
            "required": false,
            "explode": false,
            "schema": {
                "type": "array",
                "minItems": 1,
                "collectionFormat": "csv",
                "default": "lba,lbb,offres",
                "items":
                    {
                        "type": "string",
                        "enum": ["lba", "lbb", "offres"]                            
                    }
            },
            "description": "Les sources de données dans lesquelles la recherche aura lieu. Ces sources doivent être séparées par des virgules", 
            "example": "offres"
        }

    },

    "definitions":
    {
        "Error" : 
        {
            "type": "object",
            "required" : ["error"],
            "properties":
            {
                "error": 
                { 
                    "type": "string",
                    "description": "Le type générique de l'erreur",
                    "example": "wrong_parameters"
                },
                "error_messages":
                {
                    "type": "array",
                    "items" : 
                    { 
                        "type":"string"
                    },
                    "description": "Une liste d'erreurs détaillées. Ex : les erreurs de paramétrage de la requête.",
                    "example":[
                        "romes : Badly formatted rome codes. Rome code must be one letter followed by 4 digit number. ex : A1234",
                        "romes : Too many rome codes. Maximum is 9."
                    ]
                }
            }
        },

        "Formations" : 
        {
            "type": "object",
            "required" : ["results"],
            "properties":
            {
                "results": 
                { 
                    "type": "array",
                    "items":
                    {
                        "$ref": "#/definitions/Formation"
                    },
                    "description": "Un tableau de formations correspondantes aux critères"
                }
            }
        },

        "Jobs" :
        {
            "type": "object",
            "description": "Cet objet contient les types d'opportunité d'emploi demandé dans le champ <strong>source</strong> de la requête.",
            "properties":
            {
                "peJobs": 
                { 
                    "$ref": "#/definitions/Offres"
                },
                "lbaCompanies":
                {
                    "$ref": "#/definitions/Societes"
                },
                "lbbCompanies":
                {
                    "$ref": "#/definitions/Societes"
                }
            }
        },

        "Offres" : 
        {
            "type": "object",
            "required" : ["results"],
            "description": "Un tableau d'offres d'emploi en alternance correspondantes aux critères",
            "properties":
            {
                "results": 
                { 
                    "type": "array",
                    "items":
                    {
                        "$ref": "#/definitions/Offre"
                    }
                }
            }
        },

        "Societes" : 
        {
            "type": "object",
            "required" : ["results"],
            "description": "Un tableau de sociétés susceptibles de recruter en alternance correspondantes aux critères.<br />Les lba ont déjà recruté des apprentis par le passé.",
            "properties":
            {
                "results": 
                { 
                    "type": "array",
                    "items":
                    {
                        "$ref": "#/definitions/Societe"
                    }
                    
                }
            }
        },

        "Formation" : 
        {
            "type": "object",
            "required": ["title","ideaType","id"],
            "properties":
            {
                "title": 
                { 
                    "type": "string",
                    "example": "CAP Monteur en Installation Thermique",
                    "description": "Le titre de la formation"
                },
                "longTitle": 
                { 
                    "type": "string",
                    "example": "MONTEUR EN INSTALLATIONS THERMIQUES (CAP)",
                    "description": "Le titre en version longue de la formation"
                },
                "ideaType": 
                { 
                    "type": "string",
                    "value": "formation",
                    "example": "formation",
                    "description": "Le type idea d'objet, ici la seule valeur possible est 'formation'"
                },
                "id": 
                { 
                    "type": "string",
                    "example": "5e8dfad720ff3b2161269d86",
                    "description": "L'identifiant de la formation dans le catalogue de la mission apprentissage."
                },
                "diplomaLevel": 
                { 
                    "type": "string",
                    "example": "3 (CAP...)",
                    "description": "Le niveau d'étude requis ."
                },
                "onisepUrl":
                {
                    "type": "string",
                    "example": "http://www.onisep.fr/http/redirection/formation/identifiant/7872",
                    "description": "Le lien vers la description de la formation sur le site de l'ONISEP"
                },
                "contact":
                {
                    "$ref": "#/definitions/Contact"
                },
                "place":
                {
                    "description": "Le lieu de formation",
                    "$ref": "#/definitions/Place"
                },
                "company":
                {
                    "$ref": "#/definitions/Company"
                },
                "romes":
                {
                   "type": "array",
                   "items":{
                       "$ref": "#/definitions/Rome"
                   }
                }
            }
        },

        "Offre" : 
        {
            "type": "object",
            "required": ["title","ideaType","url","job"],
            "properties":
            {
                "title": 
                { 
                    "type": "string",
                    "example": "Monteur / Monteuse en chauffage (H/F)",
                    "description": "Le titre de l'offre"
                },
                "ideaType": 
                { 
                    "type": "string",
                    "value": "peJob",
                    "example": "peJob",
                    "description": "Le type d'objet, ici la seule valeur possible est 'peJob'"
                },
                "url":
                {
                    "type": "string",
                    "example": "https://candidat.pole-emploi.fr/offres/recherche/detail/106ZWVB",
                    "description": "Le lien Internet vers l'offre sur le site de Pôle emploi"
                },
                "job":
                {
                    "$ref": "#/definitions/Job"
                },
                "contact":
                {
                    "$ref": "#/definitions/Contact"
                },
                "place":
                {
                    "description": "Les coordonnées de la société proposant l'offre",
                    "$ref": "#/definitions/Place"
                },
                "company":
                {
                    "$ref": "#/definitions/Company"
                },
                "romes":
                {
                   "type": "array",
                   "items":{
                       "$ref": "#/definitions/Rome"
                   }
                }
            }
        },

        "Societe" : 
        {
            "type": "object",
            "required": ["title","ideaType","url"],
            "properties":
            {
                "title": 
                { 
                    "type": "string",
                    "example": "COMPAGNIE PARISIENNE DE CHAUFFAGE URBAIN",
                    "description": "La raison sociale de la société"
                },
                "ideaType": 
                { 
                    "type": "string",
                    "example": "lba",
                    "description": "Le type d'objet, une société peut être de type lba si elle a déjà recruté en alternance ou de type lbb si elle a recruté sur le métier recherché"
                },
                "contact":
                {
                    "$ref": "#/definitions/Contact"
                },
                "place":
                {
                    "description": "L'adresse de la société",
                    "$ref": "#/definitions/Place"
                },
                "company":
                {
                    "$ref": "#/definitions/Company"
                },
                "romes":
                {
                   "type": "array",
                   "items":{
                       "$ref": "#/definitions/Rome"
                   }
                },
                "nafs":
                {
                   "type": "array",
                   "items":{
                       "$ref": "#/definitions/Naf"
                   }
                }
            }
        },

        "Contact":
        {
            "type": "object",
            "properties":
            {
                "email":
                {
                    "type": "string",
                    "example": "contact@domaine.fr",
                    "description": "L'adresse email du contact de référence"
                },
                "name":
                {
                    "type": "string",
                    "example": "Mme Dupont",
                    "description": "Le nom du contact de référence"
                },
                "phone":
                {
                    "type": "string",
                    "example": "0X XX XX XX XX",
                    "description": "Le numéro de téléphone du contact de référence"
                },
                "info":
                {
                    "type": "string",
                    "description": "Des informations complémentaires concernant le contact de référence"
                }                      
            }            
        },

        "Place":
        {
            "type": "object",
            "description": "Un lieu. Selon le contexte il peut s'agir du lieu de formation, du lieu du centre de formation responsable de la formation ou de l'emplacement d'une entreprise offrant une opportunité d'emploi",

            "properties":
            {
                "distance":
                {
                    "type": "number",
                    "example": "3.5",
                    "description": "La distance du lieu par rapport au centre de recherche en kilomètres"
                },
                "fullAddress":
                {
                    "type": "string",
                    "example": "4 RUE DES ROSIERS PARIS 4 75004",
                    "description": "L'adresse postale complète du lieu"
                },
                "latitude":
                {
                    "type": "number",
                    "example": "48.845",
                    "description": "La latitude du lieu"
                },
                "longitude":
                {
                    "type": "string",
                    "example" : "2.3752",
                    "description": "La longitude du lieu"
                },
                "city":
                {
                    "type": "string",
                    "example": "PARIS 4",
                    "description": "La ville"
                },
                "address":
                {
                    "type": "string",
                    "example": "4 RUE DES ROSIERS",
                    "description": "L'adresse du lieu"
                },
                "cedex":
                {
                    "type": "string",
                    "example": null,
                    "description": "L'éventuel CEDEX de l'adresse"
                },
                "zipCode":
                {
                    "type": "string",
                    "example": "75004",
                    "description": "Le code postal du lieu"
                },
                "insee":
                {
                    "type": "string",
                    "example": null,
                    "description": "L'éventuel code Insee de la commune du lieu"
                }
            }      
        },

        "Job":
        {
            "type": "object",
            "description": "Les informations d'une offre d'emploi servie par l'api Offres de Pôle emploi",
            "required": ["id","description"],
            "properties":
            {
                "id":
                {
                    "type": "string",
                    "example": "106CBHM",
                    "description": "L'identifiant de l'offre"
                },

                "creationDate":
                {
                    "type": "string",
                    "example": "2020-10-01T16:43:24+02:00",
                    "description": "La date de création de l'offre sur le site de Pôle emploi"
                },
                "description":
                {
                    "type": "string",
                    "description": "La description de l'offre d'emploi"
                },
                "contractType":
                {
                    "type": "string",
                    "example" : "CDD",
                    "description": "Le type de contrat proposé"
                },
                "contractDescription":
                {
                    "type": "string",
                    "example": "Contrat à durée déterminée - 24 Mois",
                    "description": "Des informations plus détaillées sur le type de contrat proposé"
                },
                "duration":
                {
                    "type": "string",
                    "example": "39H Horaires normaux",
                    "description": "Le volume horaire de l'offre d'emploi"
                }
            }      
        },

        "Company":
        {
            "type": "object",
            "description": "Les informations de la société. Selon le contexte il peut s'agir du centre de formation ou de l'entreprise offrant une opportunité d'emploi.",
            "properties": 
            {
                "name":
                {
                    "type": "string",
                    "example": "ECOLE DE TRAVAIL ORT",
                    "description": "La raison sociale de l'entreprise"
                },
                "siret":
                {
                    "type": "string",
                    "example": "78424186100011",
                    "description": "Le numéro de SIRET de l'établissement"
                },
                "size":
                {
                    "type": "string",
                    "description": "La tranche d'effectifs de l'entreprise"
                },
                "logo":
                {
                    "type": "string",
                    "description": "L'url du logo de l'entreprise"
                },
                "description":
                {
                    "type": "string",
                    "description": "La description de l'entreprise"
                },
                "socialNetwork":
                {
                    "type": "string",
                    "description": "Un lien vers un réseau social présentant l'entreprise"
                },
                "url":
                {
                    "type": "string",
                    "description": "Un lien vers le site de l'entreprise"
                },
                "headquarter":
                {
                    "type": "object",
                    "properties":{
                        "name": {
                            "type": "string",
                            "example": "ECOLE DE TRAVAIL ORT",
                            "description": "La raison sociale du centre de formation auquel est affilié la formation"
                        },
                        "place":
                        {
                            "$ref": "#/definitions/Place"
                        }
                    }                                        
                }
            }
        },

        "Rome":
        {
            "type": "object",
            "required": ["code"],
            "properties":
            {
                "code":
                {
                    "type": "string",
                    "example": "F1603",
                    "description": "Un code ROME provenant de la nomenclature des métiers de Pôle emploi"
                },
                "label":
                {
                    "type": "string",
                    "example": "Installation d'équipements sanitaires et thermiques",
                    "description": "Le libellé correspondant au code ROME"
                }                      
            }
        },

        "Naf":
        {
            "type": "object",
            "required": ["code"],
            "properties":
            {
                "code":
                {
                    "type": "string",
                    "example": "3530Z",
                    "description": "Un code NAF (ou APE) provenant de la nomenclature des activités de l'Insee"
                },
                "label":
                {
                    "type": "string",
                    "example": "Production et distribution de vapeur et d'air conditionné",
                    "description": "Le libellé correspondant au code NAF"
                }                      
            }
        }
    }

}
  